- Removed deprecated authentication scheme
- Improved error handling, introduced error templates
- Enabled additional model properties
- Renamed `BulkComponentSPricePointAssignment` to `BulkComponentsPricePointAssignment`
- Add query parameters for list subscription group proforma invoices
- Use ProformaInvoiceStatus for list proforma invoices status query parameter
- Added missing fields in the `Allocation` model
- Added missing `404` errors for subscription components endpoints
- Renamed components price points related models to match the proper camelCase style
- Changed `includeNextProformaInvoice` param type from `String` to an enum
- Add enums for proforma invoices and invoices properties:
  - Invoice Discount Source Type
  - Invoice Discount Type
  - Proforma Invoice Discount Source Type
  - Proforma Invoice Role
  - Proforma Invoice Status
  - Tax Source Type
- Remove `ProformaTaxBreakout`, use `InvoiceTaxBreakout` instead
- Make Proforma Invoice reuse `InvoiceConsolidationLevel` and `CollectionMethod` enums instead of strings
- Remove ProformaInvoiceDiscountBreakout - use InvoiceDiscountBreakout instead
- Add missing fields for ProformaInvoiceDiscount
- Add tests for Proforma Invoices
- Merge proforma invoice and preview proforma, add kind for invoice line item
- Add kind to invoice line item
- Add apply_on_subscription_expiration to Coupon models

+++ models/replay_webhooks_request.go
@@ -6,7 +6,7 @@ import (
 
 // ReplayWebhooksRequest represents a ReplayWebhooksRequest struct.
 type ReplayWebhooksRequest struct {
-    Ids []int `json:"ids"`
+    Ids []int64 `json:"ids"`
 }
 
 // MarshalJSON implements the json.Marshaler interface for ReplayWebhooksRequest.
@@ -28,7 +28,7 @@ func (r *ReplayWebhooksRequest) toMap() map[string]any {
 // It customizes the JSON unmarshaling process for ReplayWebhooksRequest objects.
 func (r *ReplayWebhooksRequest) UnmarshalJSON(input []byte) error {
     temp := &struct {
-        Ids []int `json:"ids"`
+        Ids []int64 `json:"ids"`
     }{}
     err := json.Unmarshal(input, &temp)
     if err != nil {
